+++ src/setup.rs
@@ -5,23 +5,19 @@
 
 #[cfg(feature = "bevy_app")]
 pub use app::*;
-use bevy_ecs::{
-    system::{Commands, EntityCommands},
-    world::{EntityWorldMut, FromWorld, World},
-};
 
 #[cfg(feature = "bevy_app")]
 mod app {
     use bevy_app::{App, MainScheduleOrder, Plugin, PreUpdate};
-    use bevy_ecs::world::FromWorld;
+    use tiny_bail::prelude::*;
 
-    use crate::{next_state::NextState, schedule::StateFlush, state::State};
+    use crate::schedule::StateFlush;
 
-    use super::{insert_state, state_exists};
+    use super::*;
 
     /// A plugin that performs the required setup for [`State`] types to function:
     ///
-    /// - Adds the [`StateFlush`] schedule to the [`MainScheduleOrder`] after [`PreUpdate`].
+    /// - Adds the [`StateFlush`] schedule to the [`MainScheduleOrder`] before [`PreUpdate`].
     /// - Adds the [`bevy_state` plugin](bevy_state::app::StatesPlugin) if the
     ///   `bevy_state` feature is enabled.
     pub struct StatePlugin;
@@ -33,10 +29,11 @@ mod app {
             app.add_plugins(bevy_state::app::StatesPlugin);
 
             // Add the `StateFlush` schedule.
-            app.init_schedule(StateFlush)
+            r!(app
+                .init_schedule(StateFlush)
                 .world_mut()
-                .resource_mut::<MainScheduleOrder>()
-                .insert_after(PreUpdate, StateFlush);
+                .get_resource_mut::<MainScheduleOrder>())
+            .insert_before(PreUpdate, StateFlush);
         }
     }
 
@@ -97,6 +94,11 @@ mod app {
         fn register_state(app: &mut App);
     }
 }
+
+use bevy_ecs::{
+    system::{Commands, EntityCommands},
+    world::{EntityWorldMut, FromWorld, World},
+};
 
 use crate::{
     next_state::{NextState, TriggerStateFlush},